LayoutTests/js/dfg-arguments-unexpected-escape-expected.txt

Tests aliased uses of 'arguments' that have an unexpected escape.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS bar(42) is 84 on all iterations including after DFG tier-up.
PASS successfullyParsed is true

TEST COMPLETE
